refactor: drop __init__.py files and inline type aliases
- Remove all __init__.py files except cli (namespace packages with importlib)
- Replace NonEmptyStr with plain str across collection and extraction schemas
- Inline UnitInterval as Field(gt=0, le=1) on variance_threshold
- Move SkillMap alias to extraction/skills.py and update all imports
- Add --import-mode=importlib to pytest and explicit hatchling packages
- Remove test_empty_string test for removed NonEmptyStr constraint
