+++ pyproject.toml
@@ -47,8 +47,8 @@ viz = [
     "seaborn",
 ]
 
-[tool.hatch.build]
-sources = ["src"]
+[tool.hatch.build.targets.wheel]
+packages = ["src/chalkline"]
 
 [tool.marimo.runtime]
 pythonpath = ["./src"]
@@ -57,6 +57,7 @@ pythonpath = ["./src"]
 extraPaths = ["src"]
 
 [tool.pytest.ini_options]
+addopts    = "--import-mode=importlib"
 cache_dir  = ".cache/pytest"
 pythonpath = ["src"]
 testpaths  = ["tests"]
