--- src/chalkline/clustering/__init__.py
@@ -1,9 +1,0 @@
-"""
-Hierarchical agglomerative clustering with comparison methods.
-
-Groups PCA-reduced job postings into career families via Ward-linkage HAC
-with cophenetic validation, cluster labeling from TF-IDF centroids, and ARI
-benchmarking against the three-sector industry taxonomy. Comparison methods
-(K-Means, DBSCAN, Mean Shift, HDBSCAN) exist for the DS5230 deliverable and
-are excluded from the production pipeline.
-"""
