--- src/chalkline/collection/__init__.py
@@ -1,3 +1,0 @@
-"""
-Corpus collection from job aggregators.
-"""
