--- src/chalkline/extraction/__init__.py
@@ -1,7 +1,0 @@
-"""
-Skill extraction, lexicon normalization, and TF-IDF vectorization.
-
-Lexicon loading and normalization indices (`lexicons`, `loaders`,
-`occupations`), Aho-Corasick skill matching with surface form augmentation
-(`skills`), and TF-IDF / binary matrix construction (`vectorize`).
-"""
