--- src/chalkline/parsing/__init__.py
@@ -1,3 +1,0 @@
-"""
-Document text extraction for resumes and posting artifacts.
-"""
