--- src/chalkline/pipeline/__init__.py
@@ -1,3 +1,0 @@
-"""
-Pipeline orchestration, configuration, and shared result types.
-"""
