--- src/chalkline/reduction/__init__.py
@@ -1,7 +1,0 @@
-"""
-PCA dimensionality reduction via TruncatedSVD.
-
-Reduces the sparse TF-IDF matrix from `extraction` into a low- dimensional
-coordinate space for clustering and resume matching, with automatic
-component selection via cumulative explained variance.
-"""
