--- tests/__init__.py
@@ -1,3 +1,0 @@
-"""
-Chalkline test suite.
-"""
