--- tests/clustering/__init__.py
@@ -1,3 +1,0 @@
-"""
-Tests for hierarchical clustering and comparison methods.
-"""
