--- tests/extraction/__init__.py
@@ -1,3 +1,0 @@
-"""
-Tests for the skill extraction and normalization module.
-"""
