--- tests/parsing/__init__.py
@@ -1,3 +1,0 @@
-"""
-Tests for the document parsing module.
-"""
