--- tests/pipeline/__init__.py
@@ -1,3 +1,0 @@
-"""
-Tests for the pipeline configuration and orchestration module.
-"""
